VULNERABILITY FIX: Shared drives no longer fetch data from any address.
This commit properly fixes the file/drive metadata injection exploit
that was based on the code for adding public shared drives and fetching
their content not filtering by owner.

This was fixed by adding a function gql.ts/getSharedDriveOwner that gets
the first transaction of Entity-Type:drive and Drive-Id of the drive in question.
This function is then used by getAllMySharedDataFileTxs and getSharedPublicDrive
to filter their GQL-queries with the original drive metadata transaction's ID.

This fix was tested by using public drive bfe2ff28-6efe-4a00-8d1e-ad5f091e28a6.
An address exists with metadata transactions that use the Drive-Id of this drive,
the unpatched CLI happily downloading these files. A drive collision was also
tested with this drive - the fix appears to handle with that as well.

Though the preliminary testing was successful, some further experimenting
may be necessary to confirm that this fix does not induce unforeseen side-effects.